fix(build): integrate latest esbuild
