+++ package.json
@@ -47,7 +47,7 @@
     "buffer": "6.0.3",
     "chalk": "4.1.2",
     "chokidar": "3.5.3",
-    "esbuild": "0.14.38",
+    "esbuild": "0.14.5",
     "esbuild-register": "3.3.2",
     "eslint": "8.15.0",
     "eslint-config-prettier": "8.5.0",
