+++ packages/cli/package.json
@@ -86,7 +86,7 @@
     "checkpoint-client": "1.1.21",
     "debug": "4.3.4",
     "dotenv": "16.0.1",
-    "esbuild": "0.14.38",
+    "esbuild": "0.14.5",
     "execa": "5.1.1",
     "fast-deep-equal": "3.1.3",
     "fs-extra": "10.1.0",
