+++ packages/debug/package.json
@@ -20,7 +20,7 @@
   "devDependencies": {
     "@types/jest": "27.5.0",
     "@types/node": "12.20.51",
-    "esbuild": "0.14.38",
+    "esbuild": "0.14.5",
     "jest": "28.1.0",
     "jest-junit": "13.2.0",
     "ts-jest": "28.0.2",
@@ -38,8 +38,9 @@
   ],
   "dependencies": {
     "@types/debug": "4.1.7",
+    "debug": "4.3.4",
     "ms": "2.1.3",
     "strip-ansi": "6.0.1"
   },
-  "sideEffects": true
+  "sideEffects": false
 }
