+++ packages/engine-core/package.json
@@ -3,7 +3,6 @@
   "version": "0.0.0",
   "license": "Apache-2.0",
   "main": "dist/index.js",
-  "module": "esm/dist/index.mjs",
   "types": "dist/index.d.ts",
   "author": "Tim Suchanek <suchanek@prisma.io>",
   "homepage": "https://www.prisma.io",
@@ -23,7 +22,7 @@
     "@swc/jest": "0.2.21",
     "@types/jest": "27.5.0",
     "@types/node": "16.11.34",
-    "esbuild": "0.14.38",
+    "esbuild": "0.14.5",
     "jest": "28.1.0",
     "jest-junit": "13.2.0",
     "typescript": "4.6.4"
