+++ packages/integration-tests/package.json
@@ -3,7 +3,6 @@
   "version": "0.0.0",
   "private": true,
   "main": "dist/index.js",
-  "module": "esm/dist/index.mjs",
   "types": "dist/index.d.ts",
   "license": "Apache-2.0",
   "author": "Tim Suchanek <suchanek@prisma.io>",
@@ -23,7 +22,7 @@
     "@types/pg": "8.6.5",
     "@types/sqlite3": "3.1.8",
     "decimal.js": "10.3.1",
-    "esbuild": "0.14.38",
+    "esbuild": "0.14.5",
     "execa": "5.1.1",
     "fs-jetpack": "4.3.1",
     "jest": "28.1.0",
