+++ packages/react-prisma/package.json
@@ -2,7 +2,6 @@
   "name": "react-prisma",
   "version": "0.0.5",
   "main": "dist/index.js",
-  "module": "esm/dist/index.mjs",
   "types": "dist/index.d.ts",
   "license": "Apache-2.0",
   "author": "Tim Suchanek <suchanek@prisma.io>",
@@ -20,7 +19,7 @@
     "@prisma/client": "workspace:*",
     "@types/jest": "27.5.0",
     "@types/node": "16.11.34",
-    "esbuild": "0.14.38",
+    "esbuild": "0.14.5",
     "jest": "28.1.0",
     "jest-junit": "13.2.0",
     "react": "18.1.0",
